test(content): update content no-bounce e2e test
