+++ src/components/content/test/no-bounce/pages/root-page/root-page.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+import { IonicPage } from '../../../../../../';
+
+@IonicPage()
+@Component({
+  templateUrl: 'root-page.html'
+})
+export class RootPage {}
